requirements.txt
opencv-python
opencv-contrib-python
numpy
matplotlib
pyyaml
